Fix: crash when opening a damaged base-graphics (#11275)

+++ src/random_access_file_type.h
@@ -51,7 +51,7 @@ public:
 	uint32_t ReadDword();
 
 	void ReadBlock(void *ptr, size_t size);
-	void SkipBytes(int n);
+	void SkipBytes(size_t n);
 };
 
 #endif /* RANDOM_ACCESS_FILE_TYPE_H */
